requirements.txt
# Core Requirements
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# MeTTa/Hyperon
hyperon==0.2.2

# Google Gemini
google-generativeai==0.3.2
langchain-google-genai==0.0.14

# LangChain
langchain-core==0.1.0
langchain==0.1.0
